[SxS] Restore split tabs in the tabstrip.
Restore split tabs from session service after creating all tabs and
tabgroups. This is so, all the tabs are in the correct position and
consecutive to create the splits.

This change does not restore visual data which will be done in the next
CL.

Change-Id: I116248566ba997c3937dcdf7e31da4fe0ddbbd4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6730778
Commit-Queue: Shibalik Mohapatra <shibalik@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486659}