Handle RWHV scale override for Android
This CL makes RWHVA handle scale override. Scale override is used
during tab casting, and it modifies ScreenInfo. RWHVB takes this
into account by using GetDeviceScaleFactor, but the current code
does not handle this.

In the case that there is no override scale, then
GetDeviceScaleFactor() / view_.GetDipScale() will be exactly 1.0 so
the pixel off-by-ones that https://crrev.com/c/6169396 was fixing
will still be okay.

When override scale is set the expected behaviour is to scale the
buffer down (or up) to fit the size it would normally be shown at.

Bug: 428882066
Test: CQ, manual testing
Change-Id: I4c7b272bd715311b48fa660e2a62409fb0f69246
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6703263
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486673}