[omnibox][toolbelt] Fix sorting suggestions when input is '@'
Desired order:

1) Verbatim SWYT suggestion
2) AI Mode starter pack
3) Gemini starter pack
4) History starter pack
5) Bookmarks starter pack
6) Page starter pack
7) Tabs starter pack
8) Toolbelt

Before: We were relying on match type to identify the verbatim SWYT
suggestion. This wasn't robust because deduping could result in a
bookmark or search-history '@' search suggestion.

This CL: Robustly uses `allowed_to_be_default` to identify the default
suggestion. This is also better scalable if we one day decide to include
other defaultable suggesitons in the '@' case.

Could would be further improved if a) we wouldn't be re-implementing a
simplified sorting algorithm and would be reusing the grouping
framework; and b) we completed the autocomplete refactor and started
unioning instead of deduping matches.

Bug: 431340768
Change-Id: If0fe1b76eada75d6243b91d267009c503ded6790
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6734521
Auto-Submit: manuk hovanesian <manukh@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486720}