[site data] Reenable interactive tests
Reenabling disabled flaky tests. There are no details in the bug on the
type of flakes and it's been 2 years since the tests were disabled. No
flakes when running the bots on the CL. I'll monitor the test suite for
flakes on ChromeOS for the next week to see if they come back.

Bug: 40901060
Change-Id: I7ce46e0b28c8c8ebe1d2ec892acf2e33a1b34c15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6732839
Commit-Queue: Olesia Marukhno <olesiamarukhno@google.com>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486753}