[DevTools] Ensure CDP's ExtraInfo-events are sent for OOPIF navigations
Usually, when a navigation turns an OOPIF into an in-process iframe,
the `onRequestWillBeSentExtraInfo` and `responseReceivedExtraInfo`
events are sent via the OOPIF target.
But because of non-deterministic ordering, sometimes the navigation is
being committed before the `ExtraInfo`-events. In this case, the
`RenderFrameDevToolsAgentHost`'s destructor is called first, and the
`ExtraInfo`-events are not being sent due to no
`RenderFrameDevToolsAgentHost` existing for the (old) target of the
OOPIF.

This CL adds `GetForWithAncestorFallback()` to
`RenderFrameDevToolsAgentHost`: if no agent host is found, find
the agent host corresponding to the frame's parent.
Calling this new method ensures that the `ExtraInfo`-events do not get lost. However, they are not being sent for the OOPIF target, but for the
parent's target.

Bug: 422636456, 40255790, 40109937
Change-Id: Icf76f7a6771336d61dbac638562ece7a61500904
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6651467
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486763}