extensions: Compile `GetCurrentHost` on android too
`GetCurrentHost` was located in `extension_view_utils.cc` which has
dependency on cpp's `View` and was only compiled for desktop. We'll need
to use this util function for desktop Android as well, so we moved to it
to `extension_ui_util.cc` which is compiled on all platforms.

While we are at it, we gave it a more descriptive name.

Bug: 429498068
Change-Id: Ic52374ae2d8d94b22cfe9eb14ded5e4d27c3c6d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6733619
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Auto-Submit: Masa Fujita <massan@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Masa Fujita <massan@google.com>
Reviewed-by: Masa Fujita <massan@google.com>
Cr-Commit-Position: refs/heads/main@{#1486802}