privacy-hub: Deprecate ArcGoogleLocationServicesEnabled policy
PrivacyHub migrated location controls from ARC to ChromeOS system. This
deprecated the legacy ArcGoogleLocationServicesEnabled policy and
introduced GoogleLocationServicesEnabled as a replacement. Even though
the latter is yet to be published in Admin Panel, we need to mark the
old policy deprecated to notify users that it no longer has an effect.

PrivacyHub is Finch enabled from M128 and enabled by default in code in
M136.

Bug: 431734407
Change-Id: I1251d0be68a028f8b21509a115053f3286c3a05b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6731983
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Commit-Queue: Zauri Meshveliani <zauri@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486854}