Store SeedReaderWriter seed-info in proto instead of in a struct.
We want to move the seed and seed-related info to a dedicated file.

Store the seed data and other seed-related info in a proto object. This
proto object will be serialized and stored in a file in a future CL. For
now just use it to store all the info in memory and only persist the
seed data.

Bug: 417138763

Change-Id: Ib4300a5b40270e1ce0abfc21525bb8dc31d129be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6682031
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Commit-Queue: Ramon Cano Aparicio <rcanoaparicio@google.com>
Cr-Commit-Position: refs/heads/main@{#1486880}