Move NoVarySearchCache storage to a subdirectory
Persisting the NoVarySearchCache files to the main cache directory
caused the simple cache to thinks its index was stale, which in turn
caused it to take a lot longer to start up.

Create a "no-vary-search" subdirectory within the cache directory to
hold the snapshot.baf and journal.baj files.

Add an Init() method to NoVarySearchCacheStorageFileOperations and call
it from NoVarySearchCacheStorage before performing any other file
operations. This allows it to create the subdirectory.

Include code in Init() to move existing snapshot.baf and journal.baj
files to the subdirectory. This avoids leaving unused files on users'
disks, and also avoids losing persisted state. This is done on a
best-effort basis; even if we fail to move the old files we still
perform persistence in the "no-vary-search" subdirectory as usual.

BUG=421927600

Change-Id: Icdcb6b5492f311c7d8683d94056e83a7c16d939a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6734828
Auto-Submit: Adam Rice <ricea@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486885}