[NTPFooter] Change toggle behavior for enterprise
TL;DR: this CL updates the behavior of the footer toggle to disable the
toggle only when `EnterpriseCustomLabelForBrowser` or
`EnterpriseLogoUrlForBrowser` policies is set.

This requirement implies that the footer toggle will only be disabled
for when the customization policy is set and not for any managed browser
as in the previous requirement. The footer state will also fallback to
the user selection if a custom policy becomes unset.

The footer will listen to any changes in the footer visibility, and
hence the policies controlling the footer visibility, through the footer
controller.

For additional context, below is the behavior of the toggle following
this CL:

The toggle section is shown when at least one of the footer notice items
can be shown:

- the management notice can be shown on any NTP when the browser is managed and one of the following is true:
 - notice is force enabled by custom label/logo policy
 - user choice is enabled (i.e. `kNtpFooterVisible` pref is true) and it is not disabled by `NTPFooterManagementNoticeEnabled` policy.

-  extension attribution can be shown on extension NTP when it has not been disabled by `NTPFooterExtensionAttributionEnabled` policy.

If the toggle section is visible, its two states are as follows:

- disabled if there are custom enterprise policies for the management notice
- checked if the footer is visible.

Bug: 427772878
Change-Id: I49b7f45d8254c16913025b30671601ce559cd4e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6705273
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Paul Adedeji <pauladedeji@google.com>
Commit-Queue: Salma Elmahallawy <esalma@google.com>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486909}