Actor: Fix tab_id bug in scraper
This is due to https://crrev.com/c/6715557 which assumed the tab_id was
set on the action itself (in this case we'd look for it in
browser_action.actions[0].navigate.tab_id) but the extension differs
from Glic in this regard, it ses the tab_id on the top-level
BrowserAction.

This CL fixes the issue by using the BrowserAction.tab_id as the
fallback_tab so that, when the Action.tab_id is not found, we'll use
that.

This is a bit of a bandaid but this path is close to being replaced by
the PerformActions path which doesn't have this issue.

Bug: 431831186
Change-Id: Iaf02585ce4dbc36ee671d76512443a746c76ec79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6731369
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486922}