Make EventDispatcher accept UiStateManager
This simplifies the logic inside the EventDispatcher, which cannot exist
without a uiStateManager anyway.

this is guaranteed through the use of raw_ref instead of raw_ptr.

Future CLs will improve the construction of the ExecutionEngine and the
ActorTask so EventDispatchers can easily be injected in code.

Bug: 424495020
Change-Id: I03bac5a01f9ba54a2d3f259af59b29c8ac3d332d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6731526
Reviewed-by: Kevin Graney <kmg@google.com>
Commit-Queue: Abe Boujane <boujane@google.com>
Reviewed-by: Christine Ying <chrstne@google.com>
Cr-Commit-Position: refs/heads/main@{#1486964}