Implement a new TabInterface handle factory that maps session IDs
This change introduces a new `SessionMappedTabHandleFactory` to create
and manage handles for TabInterface, while also maintaining a mapping of
handles to session IDs.

This is currently not used, but will be inherited by TabInterface in a
follow-up CL.

The `support_handles.h` header was also updated to support deriving from
a "base" handle factory, which can be declared using a new
`DECLARE_BASE_HANDLE_FACTORY` macro.

Bug: 428721699
Change-Id: I27a73a98b4d02691753dba9018cb217ed7f2a559
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6728576
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Kaan Alsan <alsan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486987}