Remove CHECK statement in RenderWidgetCreated.
This CL removes the CHECK statement in created_widgets_ reverting to the
state prior to the change https://crrev.com/c/5961744. This CHECK causes
crashes when a new RenderWidgetHost is created with a FrameSinkId that
already exists in the `created_widgets_` map which can happen in cases
of speculative frames.

Bug: 430336826
Change-Id: I78a41ceb6143c89dcc4bcb39a27168e9f66f5c9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6731421
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Aman Verma <amanvr@google.com>
Cr-Commit-Position: refs/heads/main@{#1487015}