mac signing: Parallelize some per-distribution operations.
This may speed up the per-distribution section of signing by up to 45%.

codesign --verify and spctl --assess will now run in parallel. This is
expected to provide a speed-up up to 20% depending on I/O throughput.

The app bundle needs to be zipped prior to notarization. Previously,
the execution would wait for zipping to be complete and then start
notarization in the background (while continuing in parallel with
processing the next distribution). Now, the zipping is also done in
the background while execution moves on to the next distribution. This
could provide a speed-up of up to 25% depending on I/O vs CPU
bottlenecks.

Bug: 426019706
Change-Id: I0a054c43c2b51f56abd0a94e78a9caf337a2a216
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6699494
Commit-Queue: Mark Mentovai <mark@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1487061}