[perf] increase the experimental rate to 100% for linux-perf-trigger
The builder has been running on 50% for 2 business days + the weekend.
Increasing to 100% to get a better idea on the runtime with the current
resources.

The workload and result look ok so far. Out of the 2200+ try jobs we
have run, ~100 of them triggered Pinpoint jobs and took on average ~51
minutes to finish.

Bug: 428758873
Change-Id: Ic4670e761f22b151561cc57b442667f0f607c33f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6734809
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#1487087}