Add test to PWA Install Page Action OnClose
Following https://crrev.com/c/6569153, this CL adds tests to the OnClose
method of the migrated PWA Install Page Action.

The goal of that method is to record metrics to indicate if the IPH was
closed by a user or by Chrome, since we close the IPH when the user
chooses to install the PWA (see [1]). This is done because the function
that decides to show/not-show the IPH later reads this metric (see [2]).
The ultimate goal is to do one of the following:
- Record an "ignore" if the user closed the IPH
- Don't record an "ignore" if Chrome closed the IPH

This CL only adds test to verify that's working as intended.

I exposed RunOnIphClosedForTesting on the new implementation because
currently the Page Actions framework is responsible for triggering the
actions callbacks, so triggering that path in our tests would require
too many modifications for little benefit.

[1]
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/web_applications/web_app_dialog_utils.cc;l=312-318;drc=d856087efc39bab8ea0cdc5adda592030ed77e5d
[2]
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/web_applications/pwa_install_page_action.cc;l=242;drc=0fe11eb6b2458303d44ae35e9bbe42c6d9921a78

Bug: 421837877 Fixed: 376283433

Change-Id: Ie275b2426c55106b466ab8c0759fd1e8df5cae22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6602773
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: André Bianchessi <abianchessi@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1487098}