[Safe Browsing] Hit & Safe Browsing Report serialize functions to Util
The SafeBrowsingUI file has hundreds of lines of code contained in
private functions. As outlined in the SafeBrowsingUI refactor document,
a new util file is to be created and filled with these private functions
as well as other structs that are shared across different classes used
in SafeBrowsingUI.

Since amount of code that will be migrated to the new util file is
large, the migration will take place in steps. This CL migrates the
functions responsible for serializing the Client Safe Browsing report
and the Hit report.

Bug: 427719727
Change-Id: I57eaf3bfa50c1b49214ec1097b785a4054e2d0b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6735125
Commit-Queue: Daniel White <danieltwhite@google.com>
Reviewed-by: thefrog <thefrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1487101}