[owners] Remove lgrey@chromium.org from content/app_shim_remote_cocoa/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/lgrey@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ie2fe565c054a1aebf114e39f86cb2958259c1a87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6715507
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1487106}