Update video picture-in-picture window title animation
This change updates the window animation by making the title (origin
and favicon), together with the controls top scrim, display for 5
seconds any time the picture-in-picture window is opened.

Any interaction with the window causes the animation to revert back to
the default implementation.

Reasoning for this change available at:
go/cme-video-pip-spoof-vulnerability-by-prompt-mimicking

Bug: 397878997
Change-Id: I2daf3d5915bc04d3a729a425e9a368e44d167a34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6697443
Reviewed-by: Fr <beaufort.francois@gmail.com>
Commit-Queue: Benjamin Keen <bkeen@google.com>
Cr-Commit-Position: refs/heads/main@{#1487129}