[minizip] Fix potential OOB in unicode path extra field parsing
The code needs to compare the CRC of the "regular" filename with the
NameCRC32 value in the unicode path extra field. However, the caller may
not have passed in a szFileName buffer large enough to hold the whole
filename, or indeed any buffer at all.

To fix this, always read the full filename into a temporary buffer, and
compute the CRC on that.

This also simplifies the logic for writing the filename to szFileName,
at the expense of using a small bit of extra memory, which shouldn't be
an issue for our use cases.

Bug: 431119343
Change-Id: Ib06e1009b6e25d2d14e36858385df46d72b7bc3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6725677
Commit-Queue: Adenilson Cavalcanti <cavalcantii@chromium.org>
Auto-Submit: Hans Wennborg <hans@chromium.org>
Reviewed-by: Adenilson Cavalcanti <cavalcantii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1487152}