extensions menu: Move HasEnterpriseForcedAccess to ManagementPolicy
The calculation for whether an extension has enterprise forced access
for a profile was done in `ExtensionsMenuViewController`, but we
determined that this will be better placed in `ManagementPolicy`.
Moreover, this will enable us to use this function for Desktop Android
as well.

Bug: 429498068
Change-Id: I1beb06123e1398f6fcbcff6dd21d861c4f5bc65d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6733662
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Commit-Queue: Masa Fujita <massan@google.com>
Auto-Submit: Masa Fujita <massan@google.com>
Cr-Commit-Position: refs/heads/main@{#1487207}