extensions menu: Destroy menu when tab or profile changes
This CL makes sure that when the tab or the profile changes, the
coordinator and thus the mediator is destroyed. This makes sure that the
related components have to only consider one tab.

This is a temporary measure - in the future, we will want to provide the
ability to keep the menu open even when the tab changes.

Bug: 429498068
Change-Id: Iaa3f8e496d8bb1be747112bc8b3619ef8322d2f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6726177
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Masa Fujita <massan@google.com>
Auto-Submit: Masa Fujita <massan@google.com>
Commit-Queue: Masa Fujita <massan@google.com>
Cr-Commit-Position: refs/heads/main@{#1487208}