[Bedrock] Refactor some controllers
The following types of refactoring were done:
1. raw_ptr members were converted to raw_ref if the pointer was not
   null and its lifetime exceeded that of the class.
2. All the objects that a class depended on were explicitly passed as
   parameters, instead of querying them from the single parameter
   BrowserWindowInterface.

Converting raw_ref to pointer is a little tedious. So in case where
&member_.get() pattern was being used "multiple" times in .cc file,
helper functions were created for ease of use - e.g., GetProfile().

In DeletionDialogController, the TabStripModel::RemoveObserver call was
removed from the destructor because TabStripModelObserver already does
that job in its destructor.

For BrowserExtensionWindowController, based on offline discussions, we
decided not to apply refactoring 2.

Bug: 430452847

Change-Id: I11569a72090de4bc780784340930105625807047
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6732097
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Tom Lukaszewicz <tluk@chromium.org>
Commit-Queue: Chandranath Bhattacharyya <chbhatt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1487215}